feat: move changePossession to a new file
